+++ sdk/appconfiguration/azure-appconfiguration-provider/samples/async_key_vault_reference_provided_clients_sample.py
@@ -5,7 +5,7 @@
 # -------------------------------------------------------------------------
 
 import asyncio
-from azure.appconfiguration.provider.aio import load_provider
+from azure.appconfiguration.provider.aio import load
 from azure.appconfiguration.provider import SettingSelector, AzureAppConfigurationKeyVaultOptions
 from azure.keyvault.secrets.aio import SecretClient
 import os
@@ -22,7 +22,7 @@ async def main():
     secret_client = SecretClient(vault_url=key_vault_uri, credential=credential)
     selects = {SettingSelector(key_filter="*", label_filter="prod")}
     key_vault_options = AzureAppConfigurationKeyVaultOptions(secret_clients=[secret_client])
-    config = await load_provider(endpoint=endpoint, credential=credential, key_vault_options=key_vault_options, selects=selects)
+    config = await load(endpoint=endpoint, credential=credential, key_vault_options=key_vault_options, selects=selects)
 
     print(config["secret"])