+++ sdk/appconfiguration/azure-appconfiguration-provider/samples/key_vault_reference_sample.py
@@ -5,7 +5,7 @@
 # -------------------------------------------------------------------------
 
 from azure.appconfiguration.provider import (
-    load_provider,
+    load,
     AzureAppConfigurationKeyVaultOptions,
     SettingSelector
 )
@@ -21,6 +21,6 @@ credential = get_credential(authority)
 key_vault_options = AzureAppConfigurationKeyVaultOptions(credential=credential)
 selects = {SettingSelector(key_filter="*", label_filter="prod")}
 
-config = load_provider(endpoint=endpoint, credential=credential, key_vault_options=key_vault_options, selects=selects)
+config = load(endpoint=endpoint, credential=credential, key_vault_options=key_vault_options, selects=selects)
 
 print(config["secret"])