+++ sdk/appconfiguration/azure-appconfiguration-provider/tests/test_async_provider.py
@@ -3,7 +3,7 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # --------------------------------------------------------------------------
-from azure.appconfiguration.provider.aio import load_provider
+from azure.appconfiguration.provider.aio import load
 from azure.appconfiguration.provider import SettingSelector
 from devtools_testutils import AzureRecordedTestCase
 from devtools_testutils.aio import recorded_by_proxy_async
@@ -12,7 +12,7 @@ from async_preparers import app_config_decorator_async
 class TestAppConfigurationProvider(AzureRecordedTestCase):
 
     async def build_provider(self, connection_string, trim_prefixes=[], selects={SettingSelector(key_filter="*", label_filter="\0")}):
-        return await load_provider(connection_string=connection_string, trim_prefixes=trim_prefixes, selects=selects)
+        return await load(connection_string=connection_string, trim_prefixes=trim_prefixes, selects=selects)
 
     # method: provider_creation
     @app_config_decorator_async